adds test for fn
