+++ python/tests/data/bjs_corrections/bjs_test_output_race_and_ethnicity_state.json
@@ -89,67 +89,12 @@
     },
     {
         "state_name": "American Samoa",
-        "race_category_id": "WHITE_NH",
-        "race": "White",
+        "race_category_id": "OTHER_STANDARD_NH",
+        "race": "Unrepresented race",
         "race_includes_hispanic": false,
-        "race_and_ethnicity": "White (Non-Hispanic)",
-        "state_fips": "60",
-        "population_pct": 0.9,
-        "prison_per_100k": null,
-        "prison_pct_share": null
-    },
-    {
-        "state_name": "American Samoa",
-        "race_category_id": "BLACK_NH",
-        "race": "Black or African American",
-        "race_includes_hispanic": false,
-        "race_and_ethnicity": "Black or African American (Non-Hispanic)",
-        "state_fips": "60",
-        "population_pct": 0.02,
-        "prison_per_100k": null,
-        "prison_pct_share": null
-    },
-    {
-        "state_name": "American Samoa",
-        "race_category_id": "HISP",
-        "race": "Hispanic or Latino",
-        "race_includes_hispanic": true,
-        "race_and_ethnicity": "Hispanic or Latino",
+        "race_and_ethnicity": "Unrepresented race (Non-Hispanic)",
         "state_fips": "60",
         "population_pct": 0.1,
-        "prison_per_100k": null,
-        "prison_pct_share": null
-    },
-    {
-        "state_name": "American Samoa",
-        "race_category_id": "AIAN_NH",
-        "race": "American Indian and Alaska Native",
-        "race_includes_hispanic": false,
-        "race_and_ethnicity": "American Indian and Alaska Native (Non-Hispanic)",
-        "state_fips": "60",
-        "population_pct": null,
-        "prison_per_100k": null,
-        "prison_pct_share": null
-    },
-    {
-        "state_name": "American Samoa",
-        "race_category_id": "ASIAN_NH",
-        "race": "Asian",
-        "race_includes_hispanic": false,
-        "race_and_ethnicity": "Asian (Non-Hispanic)",
-        "state_fips": "60",
-        "population_pct": 3.6,
-        "prison_per_100k": null,
-        "prison_pct_share": null
-    },
-    {
-        "state_name": "American Samoa",
-        "race_category_id": "NHPI_NH",
-        "race": "Native Hawaiian and Pacific Islander",
-        "race_includes_hispanic": false,
-        "race_and_ethnicity": "Native Hawaiian and Pacific Islander (Non-Hispanic)",
-        "state_fips": "60",
-        "population_pct": 92.6,
         "prison_per_100k": null,
         "prison_pct_share": null
     },
@@ -166,12 +111,67 @@
     },
     {
         "state_name": "American Samoa",
-        "race_category_id": "OTHER_STANDARD_NH",
-        "race": "Unrepresented race",
+        "race_category_id": "NHPI_NH",
+        "race": "Native Hawaiian and Pacific Islander",
         "race_includes_hispanic": false,
-        "race_and_ethnicity": "Unrepresented race (Non-Hispanic)",
+        "race_and_ethnicity": "Native Hawaiian and Pacific Islander (Non-Hispanic)",
+        "state_fips": "60",
+        "population_pct": 92.6,
+        "prison_per_100k": null,
+        "prison_pct_share": null
+    },
+    {
+        "state_name": "American Samoa",
+        "race_category_id": "ASIAN_NH",
+        "race": "Asian",
+        "race_includes_hispanic": false,
+        "race_and_ethnicity": "Asian (Non-Hispanic)",
+        "state_fips": "60",
+        "population_pct": 3.6,
+        "prison_per_100k": null,
+        "prison_pct_share": null
+    },
+    {
+        "state_name": "American Samoa",
+        "race_category_id": "AIAN_NH",
+        "race": "American Indian and Alaska Native",
+        "race_includes_hispanic": false,
+        "race_and_ethnicity": "American Indian and Alaska Native (Non-Hispanic)",
+        "state_fips": "60",
+        "population_pct": null,
+        "prison_per_100k": null,
+        "prison_pct_share": null
+    },
+    {
+        "state_name": "American Samoa",
+        "race_category_id": "HISP",
+        "race": "Hispanic or Latino",
+        "race_includes_hispanic": true,
+        "race_and_ethnicity": "Hispanic or Latino",
         "state_fips": "60",
         "population_pct": 0.1,
+        "prison_per_100k": null,
+        "prison_pct_share": null
+    },
+    {
+        "state_name": "American Samoa",
+        "race_category_id": "BLACK_NH",
+        "race": "Black or African American",
+        "race_includes_hispanic": false,
+        "race_and_ethnicity": "Black or African American (Non-Hispanic)",
+        "state_fips": "60",
+        "population_pct": 0.02,
+        "prison_per_100k": null,
+        "prison_pct_share": null
+    },
+    {
+        "state_name": "American Samoa",
+        "race_category_id": "WHITE_NH",
+        "race": "White",
+        "race_includes_hispanic": false,
+        "race_and_ethnicity": "White (Non-Hispanic)",
+        "state_fips": "60",
+        "population_pct": 0.9,
         "prison_per_100k": null,
         "prison_pct_share": null
     },
